Refactor cache poisoning templates by level
