+++ src/main/java/org/sasanlabs/service/vulnerability/cachePoisoning/CachePoisoningVulnerability.java
@@ -69,7 +69,7 @@ public class CachePoisoningVulnerability {
             payload = "CACHE_POISONING_PAYLOAD_LEVEL_1_XSS")
     @VulnerableAppRequestMapping(
             value = LevelConstants.LEVEL_1,
-            htmlTemplate = "Common/CachePoisoning")
+            htmlTemplate = "LEVEL_1/CachePoisoning")
     public ResponseEntity<GenericVulnerabilityResponseBean<String>> getVulnerablePayloadLevel1(
             @RequestParam(value = BANNER_QUERY_PARAMETER, required = false) String banner,
             @RequestParam(
@@ -97,7 +97,7 @@ public class CachePoisoningVulnerability {
             payload = "CACHE_POISONING_PAYLOAD_LEVEL_2_XSS")
     @VulnerableAppRequestMapping(
             value = LevelConstants.LEVEL_2,
-            htmlTemplate = "Common/CachePoisoning")
+            htmlTemplate = "LEVEL_2/CachePoisoning")
     public ResponseEntity<GenericVulnerabilityResponseBean<String>> getVulnerablePayloadLevel2(
             @RequestParam(value = BANNER_QUERY_PARAMETER, required = false) String banner,
             @RequestParam(
@@ -125,7 +125,7 @@ public class CachePoisoningVulnerability {
             payload = "CACHE_POISONING_PAYLOAD_LEVEL_3_JS_INJECTION")
     @VulnerableAppRequestMapping(
             value = LevelConstants.LEVEL_3,
-            htmlTemplate = "Common/CachePoisoning")
+            htmlTemplate = "LEVEL_3/CachePoisoning")
     public ResponseEntity<GenericVulnerabilityResponseBean<String>> getVulnerablePayloadLevel3(
             @RequestParam(value = BANNER_QUERY_PARAMETER, required = false) String banner,
             @RequestParam(
@@ -153,7 +153,7 @@ public class CachePoisoningVulnerability {
             payload = "CACHE_POISONING_PAYLOAD_LEVEL_4")
     @VulnerableAppRequestMapping(
             value = LevelConstants.LEVEL_4,
-            htmlTemplate = "Common/CachePoisoning")
+            htmlTemplate = "LEVEL_4/CachePoisoning")
     public ResponseEntity<GenericVulnerabilityResponseBean<String>> getVulnerablePayloadLevel4(
             @RequestParam(
                             value = RESET_CACHE_QUERY_PARAMETER,
@@ -176,7 +176,7 @@ public class CachePoisoningVulnerability {
             payload = "CACHE_POISONING_PAYLOAD_LEVEL_5")
     @VulnerableAppRequestMapping(
             value = LevelConstants.LEVEL_5,
-            htmlTemplate = "Common/CachePoisoning",
+            htmlTemplate = "LEVEL_5/CachePoisoning",
             variant = Variant.SECURE)
     public ResponseEntity<GenericVulnerabilityResponseBean<String>> getSecurePayloadLevel5(
             @RequestParam(value = BANNER_QUERY_PARAMETER, required = false) String banner,
