+++ src/main/resources/static/templates/CachePoisoning/LEVEL_2/CachePoisoning.html
@@ -1,0 +1,57 @@
+<div id="cachePoisoningContainer" class="cache-poisoning-shell">
+    <h3 class="cache-poisoning-title">Cache Poisoning - Level 2</h3>
+    <p class="cache-poisoning-subtitle">
+        This level blocks only obvious script payloads. Try a harmless-looking HTML banner or a misleading status message and observe that the shared cache still replays it to the next visitor.
+    </p>
+
+    <div class="cache-poisoning-layout">
+        <section class="cache-poisoning-block">
+            <h5>Controls</h5>
+            <div class="cache-poisoning-form">
+                <div class="cache-poisoning-entry-row">
+                    <div class="cache-poisoning-field">
+                        <label class="cache-poisoning-label" for="bannerInput">Filtered banner</label>
+                        <input
+                            class="cache-poisoning-input"
+                            type="text"
+                            id="bannerInput"
+                            placeholder="&lt;b&gt;URGENT STATUS PAGE UPDATE&lt;/b&gt;" />
+                    </div>
+                    <div class="cache-poisoning-action-group cache-poisoning-action-group-inline">
+                        <button type="button" id="poisonCacheBtn">Send attacker request</button>
+                    </div>
+                </div>
+
+                <div class="cache-poisoning-actions">
+                    <div class="cache-poisoning-action-group cache-poisoning-action-group-left">
+                        <button type="button" id="resetCacheBtn" disabled>Reset Cache</button>
+                    </div>
+                    <div class="cache-poisoning-action-group cache-poisoning-action-group-right">
+                        <button type="button" id="victimRequestBtn">Send victim request</button>
+                    </div>
+                </div>
+            </div>
+        </section>
+
+        <section class="cache-poisoning-block">
+            <h5>Cache Signals</h5>
+            <div class="cache-poisoning-diagnostics">
+                <div class="diagnostic-card">
+                    <span class="diagnostic-label">Cache status</span>
+                    <code id="cacheStatus" class="cache-status-indicator cache-status-neutral">-</code>
+                </div>
+                <div class="diagnostic-card">
+                    <span class="diagnostic-label">Cache key</span>
+                    <code id="cacheKey">-</code>
+                </div>
+            </div>
+        </section>
+
+        <section class="cache-poisoning-block">
+            <h5>Observed Response</h5>
+            <div id="cachePoisoningResponse" class="cache-poisoning-output">
+                Poison the cache with a filtered banner, then send the victim request and confirm the same response is replayed from the shared cache.
+            </div>
+        </section>
+    </div>
+</div>
