+++ src/main/resources/static/templates/CachePoisoning/LEVEL_3/CachePoisoning.css
@@ -1,0 +1,1 @@
+@import url("../Common/CachePoisoning.css");
