+++ src/main/resources/static/templates/CachePoisoning/LEVEL_3/CachePoisoning.html
@@ -1,54 +1,38 @@
 <div id="cachePoisoningContainer" class="cache-poisoning-shell">
-    <h3 id="cpTitle" class="cache-poisoning-title">Cache Poisoning</h3>
-    <p id="cpSubtitle" class="cache-poisoning-subtitle"></p>
+    <h3 class="cache-poisoning-title">Cache Poisoning - Level 3</h3>
+    <p class="cache-poisoning-subtitle">
+        This level fixes the banner cache key, but still trusts <code>X-Forwarded-Host</code> when building absolute asset URLs. Use the same banner value for both requests and change only the attacker header.
+    </p>
 
     <div class="cache-poisoning-layout">
         <section class="cache-poisoning-block">
             <h5>Controls</h5>
             <div class="cache-poisoning-form">
-                <!-- Banner Input Row -->
-                <div id="bannerInputRow" class="cache-poisoning-entry-row hidden">
+                <div class="cache-poisoning-entry-row">
                     <div class="cache-poisoning-field">
-                        <label id="bannerLabel" class="cache-poisoning-label" for="bannerInput"></label>
+                        <label class="cache-poisoning-label" for="bannerInput">Shared banner key</label>
                         <input
                             class="cache-poisoning-input"
                             type="text"
                             id="bannerInput"
-                            placeholder="" />
+                            placeholder="shared-status-page" />
                     </div>
-                    <div id="poisonCacheBtnContainerInline" class="cache-poisoning-action-group cache-poisoning-action-group-inline hidden">
-                        <button type="button" id="poisonCacheBtnInline">Send attacker request</button>
+                    <div class="cache-poisoning-action-group cache-poisoning-action-group-inline">
+                        <button type="button" id="poisonCacheBtn">Send attacker request</button>
                     </div>
                 </div>
 
-                <!-- Forwarded Host Input Row -->
-                <div id="forwardedHostInputRow" class="cache-poisoning-entry-row hidden">
+                <div class="cache-poisoning-entry-row">
                     <div class="cache-poisoning-field">
-                        <label id="forwardedHostLabel" class="cache-poisoning-label" for="forwardedHostInput"></label>
+                        <label class="cache-poisoning-label" for="forwardedHostInput">Attacker forwarded host</label>
                         <input
                             class="cache-poisoning-input"
                             type="text"
                             id="forwardedHostInput"
-                            placeholder="" />
+                            placeholder="attacker.example" />
                     </div>
                 </div>
 
-                <!-- Demo User Input Row -->
-                <div id="demoUserInputRow" class="cache-poisoning-entry-row hidden">
-                    <div class="cache-poisoning-field">
-                        <label id="demoUserLabel" class="cache-poisoning-label" for="demoUserInput"></label>
-                        <input
-                            class="cache-poisoning-input"
-                            type="text"
-                            id="demoUserInput"
-                            placeholder="" />
-                    </div>
-                    <div id="poisonCacheBtnContainerDemo" class="cache-poisoning-action-group cache-poisoning-action-group-inline hidden">
-                        <button type="button" id="poisonCacheBtnDemo">Send attacker request</button>
-                    </div>
-                </div>
-
-                <!-- Global Actions -->
                 <div class="cache-poisoning-actions">
                     <div class="cache-poisoning-action-group cache-poisoning-action-group-left">
                         <button type="button" id="resetCacheBtn" disabled>Reset Cache</button>
@@ -71,11 +55,11 @@
                     <span class="diagnostic-label">Cache key</span>
                     <code id="cacheKey">-</code>
                 </div>
-                <div id="cacheControlCard" class="diagnostic-card hidden">
+                <div class="diagnostic-card">
                     <span class="diagnostic-label">Cache-Control</span>
                     <code id="cacheControl">-</code>
                 </div>
-                <div id="varyHeaderCard" class="diagnostic-card hidden">
+                <div class="diagnostic-card">
                     <span class="diagnostic-label">Vary</span>
                     <code id="varyHeader">-</code>
                 </div>
@@ -84,7 +68,9 @@
 
         <section class="cache-poisoning-block">
             <h5>Observed Response</h5>
-            <div id="cachePoisoningResponse" class="cache-poisoning-output"></div>
+            <div id="cachePoisoningResponse" class="cache-poisoning-output">
+                Send the attacker request with a forwarded host, then repeat the victim request with the same banner key and no forwarded header.
+            </div>
         </section>
     </div>
 </div>
