+++ src/main/resources/static/templates/CachePoisoning/LEVEL_4/CachePoisoning.css
@@ -1,0 +1,1 @@
+@import url("../Common/CachePoisoning.css");
