+++ src/main/resources/static/templates/CachePoisoning/LEVEL_5/CachePoisoning.html
@@ -1,0 +1,87 @@
+<div id="cachePoisoningContainer" class="cache-poisoning-shell">
+    <h3 class="cache-poisoning-title">Cache Poisoning - Level 5 (Secure)</h3>
+    <p class="cache-poisoning-subtitle">
+        This secure level ignores untrusted forwarding headers and uses a private, non-stored response for personalized content. Try sending the same malicious inputs and compare the diagnostics with the previous levels.
+    </p>
+
+    <div class="cache-poisoning-layout">
+        <section class="cache-poisoning-block">
+            <h5>Controls</h5>
+            <div class="cache-poisoning-form">
+                <div class="cache-poisoning-entry-row">
+                    <div class="cache-poisoning-field">
+                        <label class="cache-poisoning-label" for="bannerInput">Optional banner preview</label>
+                        <input
+                            class="cache-poisoning-input"
+                            type="text"
+                            id="bannerInput"
+                            placeholder="&lt;b&gt;OWNED&lt;/b&gt;" />
+                    </div>
+                </div>
+
+                <div class="cache-poisoning-entry-row">
+                    <div class="cache-poisoning-field">
+                        <label class="cache-poisoning-label" for="forwardedHostInput">Optional forwarded host</label>
+                        <input
+                            class="cache-poisoning-input"
+                            type="text"
+                            id="forwardedHostInput"
+                            placeholder="attacker.example" />
+                    </div>
+                </div>
+
+                <div class="cache-poisoning-entry-row">
+                    <div class="cache-poisoning-field">
+                        <label class="cache-poisoning-label" for="demoUserInput">Personalized demo_user</label>
+                        <input
+                            class="cache-poisoning-input"
+                            type="text"
+                            id="demoUserInput"
+                            placeholder="admin" />
+                    </div>
+                    <div class="cache-poisoning-action-group cache-poisoning-action-group-inline">
+                        <button type="button" id="poisonCacheBtn">Send attacker request</button>
+                    </div>
+                </div>
+
+                <div class="cache-poisoning-actions">
+                    <div class="cache-poisoning-action-group cache-poisoning-action-group-left">
+                        <button type="button" id="resetCacheBtn" disabled>Reset Cache</button>
+                    </div>
+                    <div class="cache-poisoning-action-group cache-poisoning-action-group-right">
+                        <button type="button" id="victimRequestBtn">Send guest request</button>
+                    </div>
+                </div>
+            </div>
+        </section>
+
+        <section class="cache-poisoning-block">
+            <h5>Cache Signals</h5>
+            <div class="cache-poisoning-diagnostics">
+                <div class="diagnostic-card">
+                    <span class="diagnostic-label">Cache status</span>
+                    <code id="cacheStatus" class="cache-status-indicator cache-status-neutral">-</code>
+                </div>
+                <div class="diagnostic-card">
+                    <span class="diagnostic-label">Cache key</span>
+                    <code id="cacheKey">-</code>
+                </div>
+                <div class="diagnostic-card">
+                    <span class="diagnostic-label">Cache-Control</span>
+                    <code id="cacheControl">-</code>
+                </div>
+                <div class="diagnostic-card">
+                    <span class="diagnostic-label">Vary</span>
+                    <code id="varyHeader">-</code>
+                </div>
+            </div>
+        </section>
+
+        <section class="cache-poisoning-block">
+            <h5>Observed Response</h5>
+            <div id="cachePoisoningResponse" class="cache-poisoning-output">
+                Send a personalized request first, then clear the cookie with a guest request. The response should stay fresh and should not reuse attacker-controlled forwarding data.
+            </div>
+        </section>
+    </div>
+</div>
